docs: add api-explorer page
