+++ packages/vuetifyjs.com/src/data/pages/components/ApiExplorer.json
@@ -1,0 +1,14 @@
+{
+  "title": "header",
+  "titleText": "headerText",
+  "children": [
+    {
+      "type": "section",
+      "children": [
+        {
+          "type": "page-api-explorer"
+        }
+      ]
+    }
+  ]
+}
